+++ rouziclib/general/mouse_struct.h
@@ -10,8 +10,8 @@ typedef struct
 
 typedef struct
 {
-	int quick;	// flag when a click is released as soon as it's pressed
-	int block;	// flag when a click is to be ignored after window focus regain
+	int8_t quick;	// flag when a click is released as soon as it's pressed
+	int8_t block;	// flag when a click is to be ignored after window focus regain
 } mousebut_flags_t;
 
 typedef struct
@@ -33,11 +33,11 @@ enum
 typedef struct
 {
 	xy_t a, u, d, u_stored, prev_u;
-	int window_focus_flag, mouse_focus_flag, window_minimised_flag, zoom_flag, zoom_scroll_freeze, showcursor, warp, warp_prev, warp_if_move, zoom_allowed, discard_warp_first_move;
-	int mod_key[mouse_mod_count];
+	int8_t window_focus_flag, mouse_focus_flag, window_minimised_flag, zoom_flag, zoom_scroll_freeze, showcursor, warp, warp_prev, warp_if_move, zoom_allowed, discard_warp_first_move;
+	int8_t mod_key[mouse_mod_count];
 	mousebut_t b;
 	mouse_ctrl_id_t *ctrl_id;
-	int key_state[RL_NUM_SCANCODES], key_quick[RL_NUM_SCANCODES];	// keyboard state array, -2 newly up, -1 up, 0 = initial, 1 = down, 2 = newly down, 3 = repeated down event
+	int8_t key_state[RL_NUM_SCANCODES], key_quick[RL_NUM_SCANCODES];	// keyboard state array, -2 newly up, -1 up, 0 = initial, 1 = down, 2 = newly down, 3 = repeated down event
 	#ifdef MOUSE_LOG
 	int frame_index;
 	buffer_t log;
